requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
